third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGRadialGradientElement-svgdom-gradientUnits-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'gradientUnits' property of the SVGRadialGradientElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS radialGradientElement.gradientUnits.baseVal is SVGUnitTypes.SVG_UNIT_TYPE_OBJECTBOUNDINGBOX
PASS radialGradientElement.gradientUnits.baseVal is SVGUnitTypes.SVG_UNIT_TYPE_USERSPACEONUSE
PASS successfullyParsed is true

TEST COMPLETE
